fix(http-runtime): r2 — shared dial module, plaintext-test resolver, candidate cap, locked design
Address quality-mgr r1 findings on PR #1142:
- move dial/cache code into peer_dial.rs shared by both wire-security modes
- plaintext-test direct reqwest client now resolves through OrderedPeerResolver
  (scope-less link-local IPv6 dropped, IPv4 first), closing the unfiltered path
- cap dial candidates at 4 so the per-attempt budget share stays usable
- outer connect guard gets DIAL_REPORT_GRACE so per-address causes are reported
- DirectPeerTcpConnector::new reuses direct_peer_authority()
- address cache TTL lives in PeerPoolConfig (default 5 min, non-zero)
- cache key normalizes ASCII case and strips trailing .local so rand-m5 and
  rand-m5.local share one entry

Lock the design: REQ-CORE-TRANSPORT-002E and ADR-060 record the dial order,
budget split, cache semantics, dual-mode guarantee, and diagnostics contract.
