+++ docs/adr/INDEX.md
@@ -66,6 +66,7 @@ crate-local ADR records that remain embedded in crate architecture documents.
 - [ADR-057 — Peer-Write Redial and Delivery-Attempt Invariant](./ADR-057-peer-write-redial-and-delivery-attempt-invariant.md)
 - [ADR-058 — Herdr Local Steer Backend Contract](./ADR-058-herdr-local-steer-backend-contract.md)
 - [ADR-059 — Async Mailbox-Read Concurrency And State Handoff](./ADR-059-async-mailbox-read-concurrency.md)
+- [ADR-060 — Peer Dial Order And Address Cache](./ADR-060-peer-dial-and-address-cache.md)
 
 ## Extracted Crate-Local ADRs
